refactor: errorMessage 레이아웃 제거 및 Providers 위치 변경
